pass if no tags to remove

+++ code/JDDB/jddb/processor/shot.py
@@ -175,7 +175,8 @@ class Shot(object):
             existing_tags = self.__original_tags
             tags_to_remove = [r_tag for r_tag in existing_tags if r_tag not in self.tags]
 
-            self.file_repo.remove_data(self.shot_no, tags_to_remove)
+            if len(tags_to_remove):
+                self.file_repo.remove_data(self.shot_no, tags_to_remove)
             data_dict = dict()
 
             for tag, signal in self.__new_signals.items():
